update
Quiz currently contains 3 quesions. If user selects an option, a button will dynamically render underneath. The button will either display Next or Results depending on whether or not the question has been answered and whether or not the question is the last question in the bank. Next step is to implement score and display score in the results.